% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{course.sources}
\alias{course.sources}
\title{Source of Courses}
\description{
A named list. The name is the course name in the records. The value is another list,
where each list is the object returned from applying \code{yaml::yaml.load_file} to the course sources.
}
