% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{get.progress.table}
\alias{get.progress.table}
\title{Extract the student progress from the records}
\usage{
get.progress.table(records, key = NULL)
}
\arguments{
\item{records}{The data returned from \code{\link{get.records}}.}

\item{key}{The key of the google spreadsheet of the project.
If the \code{key} is not specified, then only students and courses recorded in the \code{records} are reported.}
}
\value{
A data.frame whose column is:
\itemize{
 \item{student} {The classroom account of the student.}
 \item{course} {The tracked course.}
 \item{status} {Indicates the progress of the student at the course. "O" : The student finished the course. "P": The student entered but not finished the course. "X": The student has not entered the course yet.}
 \item{finish.time} {The first time to complete.}
 \item{duration} {The time student spent on the course before the first accomplishment.}
 \item{comment} {Additional information from the records.}
}
}
\examples{
\dontrun{
The teacher could convert the result to wild table as follow:
library(reshape2)
# login first
records <- get.records()
result <- get.progress.table(records)
dcast(result, student ~ course, value.var = "status")
}
}
