% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{login}
\alias{login}
\title{Login to the DataScienceAndR Service}
\usage{
login(account)
}
\arguments{
\item{account}{The registered account of the teacher}
}
\value{
NULL. It will raise an error if login is failed.
}
\examples{
\dontrun{
login("wush")
# input the password

# get records of all associated students
records <- get.records()
}
}
