% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{skipped.analysis}
\alias{skipped.analysis}
\title{Analyze whether the student skipped the questions or not}
\usage{
skipped.analysis(records, students = NULL, lessons = DSR.lessons,
  course.source = course.sources)
}
\arguments{
\item{records}{The data.frame of the raw records. This should be the output of the function
\code{get.records}.}

\item{students}{The user_id to be analyzed. If it is \code{NULL}, then all students in the \code{records} will
be analyzed.}

\item{lessons}{The character vector of the courses to be analyzed.}

\item{course.source}{named list. The name is the course name.
The value is the root of the swirl course directory.}
}
\value{
data.frame with following columns: \itemize{
 \item course The course name.
 \item question.no The number of the question
 \item user_id The id of the student
 \item skipped Whether the question is skipped or not
 \item source.position integer. The index of which part of source found in \code{course.source}
}
Note that the students might do the question multiple times, so
there might be multiple rows for the specific course, question.no and user_id.
Each row represents each trials of the student at the specific course and question.
}
\examples{
\dontrun{
login("wush")
records <- get.records()
skipped.result <- skipped.analysis(records)
library(dplyr)
group_by(skipped.result, course, question.no, user_id) \%>\%
 summarise(is.skipped = all(skipped)) \%>\% # if the student does not skip once, then I recognize that he/she did the question.
 group_by(course, question.no) \%>\%
 summarise(skip.ratio = mean(is.skipped)) \%>\%
 arrange(desc(skip.ratio))
}
}
